Wrote more tests

+++ cdh/source/telemetry_pack.c
@@ -5,7 +5,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-typedef void (*telemetry_pack_func_t)(telemetry_data_t *, uint8_t *, size_t *);
+typedef void (*telemetry_pack_func_t)(telemetry_data_t *, uint8_t *, uint32_t *);
 
 static const telemetry_pack_func_t telemPackFns[] = {
     [TELEM_CC1120_TEMP] = packCC1120Temp,
@@ -36,10 +36,10 @@ static const telemetry_pack_func_t telemPackFns[] = {
     [TELEM_PONG] = packPong,
 };
 
-static void packTelemetryId(telemetry_data_id_t id, uint8_t *buffer, size_t *offset);
-static void packTelemetryTimestamp(uint32_t timestamp, uint8_t *buffer, size_t *offset);
+static void packTelemetryId(telemetry_data_id_t id, uint8_t *buffer, uint32_t *offset);
+static void packTelemetryTimestamp(uint32_t timestamp, uint8_t *buffer, uint32_t *offset);
 
-obc_error_code_t packTelemetry(telemetry_data_t *data, uint8_t *buffer, size_t len, size_t *numPacked) {
+obc_error_code_t packTelemetry(telemetry_data_t *data, uint8_t *buffer, size_t len, uint32_t *numPacked) {
   if (data == NULL || buffer == NULL || numPacked == NULL) {
     return OBC_ERR_CODE_INVALID_ARG;
   }
@@ -52,7 +52,7 @@ obc_error_code_t packTelemetry(telemetry_data_t *data, uint8_t *buffer, size_t l
     return OBC_ERR_CODE_INVALID_ARG;
   }
 
-  size_t offset = 0;
+  uint32_t offset = 0;
 
   // Pack the telemetry ID
   packTelemetryId(data->id, buffer, &offset);
@@ -68,110 +68,114 @@ obc_error_code_t packTelemetry(telemetry_data_t *data, uint8_t *buffer, size_t l
   return OBC_ERR_CODE_SUCCESS;
 }
 
-static void packTelemetryId(telemetry_data_id_t id, uint8_t *buffer, size_t *offset) { packUint8(id, buffer, offset); }
+static void packTelemetryId(telemetry_data_id_t id, uint8_t *buffer, uint32_t *offset) {
+  packUint8(id, buffer, offset);
+}
 
-static void packTelemetryTimestamp(uint32_t timestamp, uint8_t *buffer, size_t *offset) {
+static void packTelemetryTimestamp(uint32_t timestamp, uint8_t *buffer, uint32_t *offset) {
   packUint32(timestamp, buffer, offset);
 }
 
-void packCC1120Temp(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packCC1120Temp(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->cc1120Temp, buffer, offset);
 }
 
-void packCommsCustomTransceiverTemp(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packCommsCustomTransceiverTemp(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->commsCustomTransceiverTemp, buffer, offset);
 }
 
-void packObcTemp(telemetry_data_t *data, uint8_t *buffer, size_t *offset) { packFloat(data->obcTemp, buffer, offset); }
+void packObcTemp(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
+  packFloat(data->obcTemp, buffer, offset);
+}
 
-void packAdcsMagBoardTemp(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packAdcsMagBoardTemp(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->adcsMagBoardTemp, buffer, offset);
 }
 
-void packAdcsSensorBoardTemp(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packAdcsSensorBoardTemp(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->adcsSensorBoardTemp, buffer, offset);
 }
 
-void packEpsBoardTemp(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packEpsBoardTemp(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->epsBoardTemp, buffer, offset);
 }
 
-void packSolarPanel1Temp(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packSolarPanel1Temp(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->solarPanel1Temp, buffer, offset);
 }
 
-void packSolarPanel2Temp(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packSolarPanel2Temp(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->solarPanel2Temp, buffer, offset);
 }
 
-void packSolarPanel3Temp(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packSolarPanel3Temp(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->solarPanel3Temp, buffer, offset);
 }
 
-void packSolarPanel4Temp(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packSolarPanel4Temp(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->solarPanel4Temp, buffer, offset);
 }
 
-void packEpsComms5vCurrent(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packEpsComms5vCurrent(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->epsComms5vCurrent, buffer, offset);
 }
 
-void packEpsComms3v3Current(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packEpsComms3v3Current(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->epsComms3v3Current, buffer, offset);
 }
 
-void packEpsMagnetorquer8vCurrent(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packEpsMagnetorquer8vCurrent(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->epsMagnetorquer8vCurrent, buffer, offset);
 }
 
-void packEpsAdcs5vCurrent(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packEpsAdcs5vCurrent(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->epsAdcs5vCurrent, buffer, offset);
 }
 
-void packEpsAdcs3v3Current(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packEpsAdcs3v3Current(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->epsAdcs3v3Current, buffer, offset);
 }
 
-void packEpsObc3v3Current(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packEpsObc3v3Current(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->epsObc3v3Current, buffer, offset);
 }
 
-void packEpsComms5vVoltage(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packEpsComms5vVoltage(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->epsComms5vVoltage, buffer, offset);
 }
 
-void packEpsComms3v3Voltage(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packEpsComms3v3Voltage(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->epsComms3v3Voltage, buffer, offset);
 }
 
-void packEpsMagnetorquer8vVoltage(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packEpsMagnetorquer8vVoltage(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->epsMagnetorquer8vVoltage, buffer, offset);
 }
 
-void packEpsAdcs5vVoltage(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packEpsAdcs5vVoltage(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->epsAdcs5vVoltage, buffer, offset);
 }
 
-void packEpsAdcs3v3Voltage(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packEpsAdcs3v3Voltage(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->epsAdcs3v3Voltage, buffer, offset);
 }
 
-void packEpsObc3v3Voltage(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packEpsObc3v3Voltage(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packFloat(data->epsObc3v3Voltage, buffer, offset);
 }
 
-void packObcState(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packObcState(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packUint8(data->obcState, buffer, offset);
 }
 
-void packEpsState(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packEpsState(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packUint8(data->epsState, buffer, offset);
 }
 
-void packNumCspPacketsRcvd(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packNumCspPacketsRcvd(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   packUint32(data->numCspPacketsRcvd, buffer, offset);
 }
 
-void packPong(telemetry_data_t *data, uint8_t *buffer, size_t *offset) {
+void packPong(telemetry_data_t *data, uint8_t *buffer, uint32_t *offset) {
   // Nothing to pack
 }
