+++ common/include/obc_pack_utils.h
@@ -15,7 +15,7 @@ extern "C" {
  * @param buffer  The buffer to pack the value into.
  * @param offset  A pointer to the offset within the buffer to pack the value at.
  */
-void packUint8(uint8_t value, uint8_t* buffer, size_t* offset);
+void packUint8(uint8_t value, uint8_t* buffer, uint32_t* offset);
 
 /**
  * Packs an unsigned 16-bit integer into a buffer at the given offset.
@@ -24,7 +24,7 @@ void packUint8(uint8_t value, uint8_t* buffer, size_t* offset);
  * @param buffer  The buffer to pack the value into.
  * @param offset  A pointer to the offset within the buffer to pack the value at.
  */
-void packUint16(uint16_t value, uint8_t* buffer, size_t* offset);
+void packUint16(uint16_t value, uint8_t* buffer, uint32_t* offset);
 
 /**
  * Packs an unsigned 32-bit integer into a buffer at the given offset.
@@ -33,7 +33,7 @@ void packUint16(uint16_t value, uint8_t* buffer, size_t* offset);
  * @param buffer  The buffer to pack the value into.
  * @param offset  A pointer to the offset within the buffer to pack the value at.
  */
-void packUint32(uint32_t value, uint8_t* buffer, size_t* offset);
+void packUint32(uint32_t value, uint8_t* buffer, uint32_t* offset);
 
 /**
  * Packs a signed 8-bit integer into a buffer at the given offset.
@@ -42,7 +42,7 @@ void packUint32(uint32_t value, uint8_t* buffer, size_t* offset);
  * @param buffer  The buffer to pack the value into.
  * @param offset  A pointer to the offset within the buffer to pack the value at.
  */
-void packInt8(int8_t value, uint8_t* buffer, size_t* offset);
+void packInt8(int8_t value, uint8_t* buffer, uint32_t* offset);
 
 /**
  * Packs a signed 16-bit integer into a buffer at the given offset.
@@ -51,7 +51,7 @@ void packInt8(int8_t value, uint8_t* buffer, size_t* offset);
  * @param buffer  The buffer to pack the value into.
  * @param offset  A pointer to the offset within the buffer to pack the value at.
  */
-void packInt16(int16_t value, uint8_t* buffer, size_t* offset);
+void packInt16(int16_t value, uint8_t* buffer, uint32_t* offset);
 
 /**
  * Packs a signed 32-bit integer into a buffer at the given offset.
@@ -60,7 +60,7 @@ void packInt16(int16_t value, uint8_t* buffer, size_t* offset);
  * @param buffer  The buffer to pack the value into.
  * @param offset  A pointer to the offset within the buffer to pack the value at.
  */
-void packInt32(int32_t value, uint8_t* buffer, size_t* offset);
+void packInt32(int32_t value, uint8_t* buffer, uint32_t* offset);
 
 /**
  * Packs a floating point value into a buffer at the given offset.
@@ -69,7 +69,7 @@ void packInt32(int32_t value, uint8_t* buffer, size_t* offset);
  * @param buffer  The buffer to pack the value into.
  * @param offset  A pointer to the offset within the buffer to pack the value at.
  */
-void packFloat(float value, uint8_t* buffer, size_t* offset);
+void packFloat(float value, uint8_t* buffer, uint32_t* offset);
 
 #ifdef __cplusplus
 }
